fix: brush_mask RLE 인코딩 언더플로우로 인한 데이터 유실 수정 (BUG-002)
rle_encode()의 np.diff가 flat의 uint8 dtype을 그대로 유지해 1->0 하강
엣지(-1)가 부호없는 정수 언더플로우로 255가 되어 ends가 항상 비고
rle_encode가 항상 빈 문자열을 반환하던 문제 수정. np.diff 호출 전
flat을 int8로 캐스팅해 해결. 기존 n = min(len(starts), len(ends))
경쟁조건 방어 코드는 유지. polygon 타입은 이 함수를 타지 않아 영향 없음.

+++ app/core/annotation_store.py
@@ -210,7 +210,9 @@ def rle_encode(mask: np.ndarray) -> str:
     flat = (flat != 0).view(np.uint8)
     if not flat.any():
         return ""
-    diff   = np.diff(flat, prepend=np.uint8(0), append=np.uint8(0))
+    # int8로 캐스팅 후 diff — uint8 그대로 diff하면 1→0 하강엣지(-1)가
+    # 부호없는 정수 언더플로우로 255가 되어 diff==-1 매칭이 항상 실패한다 (BUG-002).
+    diff   = np.diff(flat.astype(np.int8), prepend=np.int8(0), append=np.int8(0))
     starts = np.where(diff == 1)[0]
     ends   = np.where(diff == -1)[0]
     # 경쟁 조건으로 전환 수가 불일치하는 경우 방어 처리
